Move `parallel_state.py` to the `distributed` folder
